added custom grid component
