+++ stories/utils/grid.tsx
@@ -1,0 +1,32 @@
+const getFaqDetails = () => {
+  const faq = [
+    {
+      question: ('question1'),
+      ans: ('ans1'),
+    },
+    {
+      question: ('question2'),
+      ans: ('ans2'),
+    },
+    {
+      question: ('question3'),
+      ans: ('ans3'),
+    },
+    {
+      question: ('question4'),
+      ans: ('ans4'),
+    },
+    {
+      question: ('question5'),
+      ans: ('ans5'),
+    },
+    {
+      question: ('question6'),
+      ans: ('ans6'),
+    }
+    ];
+
+  return (faq);
+};
+
+export default getFaqDetails;
